BROKEN: allow prefix in monad (& m-seq); hit a bit of a wall
So I'm actually a bit stuck with the syntax I'm going for here.

Here's the issue. I want to be able to namespace my monadic operations,
which means I need the `monad` macro to know it needs to output
`:m/bind` instead of `:bind`. That's relatviely easy to do, and working
in the current state of this code.

At least, as long as the provided keyword representing the namespace is
a literal keyword. So, this works:

```clojure
(monad :m
  a :<< some-expr
  b :<< other-expr
  [:m/pure (+ a b)])
```

with the glue being `:m/bind`. But, I also want to be able to write
generic monadic operations, like `m-seq`. Which means that now `m-seq`
needs to know about the prefix. But `m-seq` is a function, and I'd
prefer if it could remain that way, because macros are bad (as we see
here, they tend to be contagious).

But that means I'm ultimately reduced to writing something like:

```clojure
(let [prefix :m]
  (monad prefix
    a :<< expr
    ...))
```

and that no longer works, because now when `monad` looks at its first
argument it no longer sees a keyword.

Because the syntax around `monad` is so loose, I can't think of a good
way for it to detect that the first argument is meant as a prefix in the
general case where it is not a literal keyword. We could go down the
path of making it a mandatory argument, but I don't really like that:
people should be able to still use unqualified keywords if they like.

We could give it more structure, for example, define the syntax as:

```clojure
(monad [:prefix :m]
  a :<< expr
  b :<< other-expr
  ...)
```

and that would work, but it feels icky. Also we'd probably want to make
it `{:prefix :m}` instead, in case we come up with other options in the
future. But then we again have a map that needs to be literal, and
that's annoying.

The core issue is that `monad` takes all of its arguments as a flat
list, and "counts" them to find the pieces. And a single expression is a
valid piece, so there's no way for us to make a distinction between a
single piece meant as a monadic operation and a single piece meant as
the prefix.

So I'm changing syntax again, I think. Having used monad a bit, I kinda
like `:<<`, but it's not very Clojure-y. I think I'm going to move back
a bit towards `mdo`, but with a tweak. Stay tuned in next PR.

To be clear, what's broken in this PR is, specifically, `m-seq` in the
presence of a keyword, as it is unable to forward it through to nested
`monad` calls.